Add IPv6 support
Previously, if `ansible_host` was a hostname, only the 1st IP address was used
and assumed to be an IPv4 address. This lead to a few issues:

- the Go docs for `net.LookupIP()` do not specify that the 1st entry will
  be an IPv4 address, so the wrong CIDR could have been constructed
  (e.g. an IPv6 address using a `/32` CIDR)

- only one IP address was used and all the others were disregarded

We now process all IP addresses of all types (IPv4 and IPv6) and do not make
assumptions about the order in which they appear.